Remove CompositeState
CompositeState is an inherintly complex problem we don't want to solve
right now as we have no consumers of it. Some challenges:

- We may want to emit an argument for each composed state to the
subscribe which would require overriding the publish and subscribe
methods of State
- We may want to allow you to map or merge the states together into a
single object rather than getting each one separately (for example what
if you want to do a calculation based on multiple states)
- Should we wait to emit until all states have emited? Or emit as soon
as the first one emits? Should we emit initially?